Release a new version under MIT instead of GPLv3 (#33)
When I first wrote this I never gave it enough thinking, I wanted people to use it, and I wanted their improvements of it to be released back to the public, that's why I chose GPLv3, but I didn't want to restrict people from using it in their non-GPL work. I didn't really understand that part. With this I'm releasing a new version under the MIT license, all the future versions will be under MIT as well.